Change Timestamp + Add Lifecycle to LikedComment + Return on login if not login actions

+++ src/Controller/CommentController.php
@@ -17,36 +17,56 @@ class CommentController extends AbstractController
     #[Route('/{id}', name: 'app_comment_remove')]
     public function remove(Comment $comment, CommentRepository $commentRepository, Request $request): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $id = $request->query->get('post');
         $commentRepository->remove($comment, true);
         return $this->redirectToRoute('app_post_show', ['id' => $id]);
     }
 
     #[Route('/{id}/like', name: 'app_comment_like')]
-    public function like(Comment $comment, LikedCommentRepository $likedCommentRepository): Response
+    public function like(Comment $comment, LikedCommentRepository $likedCommentRepository, Request $request): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        $id = $request->query->get('post');
+
+        if ($this->getUser() !== $comment->getUser()) {
+            return $this->redirectToRoute('app_post_show', ['id' => $id]);
+        }
+
         $likedComment = new LikedComment();
 
         if ($likedCommentRepository->findOneBy(['comment'=>$comment, 'user'=>$this->getUser()])) {
-            return $this->redirectToRoute('app_post_show', ['id' => $comment->getId()]);
+            return $this->redirectToRoute('app_post_show', ['id' => $id]);
         }
 
-        $likedComment->setComment($comment);
-        $likedComment->setUser($this->getUser());
+        $likedComment->setComment($comment)
+            ->setUser($this->getUser());
         $likedCommentRepository->save($likedComment, true);
-        return $this->redirectToRoute('app_post_show', ['id' => $comment->getId()]);
+        return $this->redirectToRoute('app_post_show', ['id' => $id]);
     }
 
     #[Route('/{id}/dislike', name: 'app_comment_dislike')]
-    public function dislike(Comment $comment, LikedCommentRepository $likedCommentRepository): Response
+    public function dislike(Comment $comment, LikedCommentRepository $likedCommentRepository, Request $request): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        $id = $request->query->get('post');
+
         if (!$likedCommentRepository->findOneBy(['comment'=>$comment, 'user'=>$this->getUser()])) {
-            return $this->redirectToRoute('app_post_show', ['id' => $comment->getId()]);
+            return $this->redirectToRoute('app_post_show', ['id' => $id]);
         }
 
         $likedComment = $likedCommentRepository->findOneBy(['comment'=>$comment, 'user'=>$this->getUser()]);
         $likedCommentRepository->remove($likedComment, true);
-        return $this->redirectToRoute('app_post_show', ['id' => $comment->getId()]);
+        return $this->redirectToRoute('app_post_show', ['id' => $id]);
     }
 
 }
